ocr-service/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==10.8.0
PyMuPDF==1.25.1
pytesseract==0.3.13
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic==2.10.2
pydantic-settings==2.6.1